server

+++ platform/Linux.c
@@ -94,15 +94,15 @@ int CW_Platform_Socket(bool isStream)
     else
     {
         int sd = socket(AF_INET, SOCK_DGRAM, IPPROTO_UDP);
-
+/*
         int flags = fcntl(sd, F_GETFL, 0);
 
         if (flags >= 0)
         {
-            flags |= (unsigned int)O_NONBLOCK;
+            flags |= O_NONBLOCK;
             fcntl(sd, F_SETFL, (int)flags);
         }
-
+*/
         return sd;
     }
 }
